requirements.txt
addict==2.2.1
aioredis==1.3.1
appnope==0.1.0
asgiref==3.2.3
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
autopep8==1.4.4
backcall==0.1.0
certifi==2019.9.11
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
decorator==4.4.0
Django==2.2.4
django-cors-headers==3.1.1
django-extensions==2.2.1
django-filter==2.2.0
django-jalali-date==0.3.1
django-location-field==2.1.0
django-rest-swagger==2.2.0
django-stubs
djangorestframework==3.10.2
djangorestframework-stubs
drf-yasg==1.17.1
furl==2.1.0
graphviz==0.12
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflection==0.3.1
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jdatetime==3.6.2
jedi==0.15.1
Jinja2==2.10.1
kavenegar==1.1.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mypy
mypy-extensions==0.4.3
openapi-codec==1.3.2
orderedmultidict==1.0.1
packaging==19.2
parso==0.5.1
persiantools==1.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydotplus==2.0.2
Pygments==2.4.2
PyHamcrest==2.0.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.2
pytz==2019.2
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
tzlocal==2.1
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
zope.interface==4.7.1
